Defer import of unittest
